Remove templating from T*KMeansScans and use abstract TClusters instead
